site/requirements.txt
django-locksmith
django-storages
markdown
docutils
